Further tidy cached tree hash
